mm/gup: track FOLL_LONGTERM pins in addition to FOLL_PIN pins
When pages are pinned for long term DMA, as indicated by the caller
passing FOLL_LONGTERM in to any of the pin_user_pages*() APIs, then
those pages need to be tracked, in order to make accurate decisions
later during memory reclaim, file writeback, and other critical places.

Fortunately, such pages also do not need to be on an LRU list while
pinned, because then cannot be managed normally (they remain pinned and
resident). And that means that the page->lru.[next,prev] fields are
available for tracking.

This patch removes pages from their LRU during a FOLL_LONGTERM pinning
call, and uses the resulting page->lru fields to track how many callers
have pinned each page, and to identify "pinned" vs. "on an LRU list".
Ultimately, the point of all this is to return true or false from these
new functions:

    folio_is_longterm_pinned()
    page_is_longterm_pinned()

This is achieved by maintaining both a count of long term pinners, and a
separate is_longterm_pinned flag. That's because a count may have the
same value as a pointer, but a flag such as 0x2 will not. The separate
flag field allows easily determining if a page is on an LRU, or is long
term pinned.

In order to keep long term pages from being prematurely put back onto an
LRU list, folio_evictable() is changed so that, in addition to rejecting
mlocked and otherwise unevicatable folios, it also rejects long term
pinned folios.

Cc: Jan Kara <jack@suse.cz>
Cc: Christoph Hellwig <hch@infradead.org>
Cc: Matthew Wilcox (Oracle) <willy@infradead.org>
Signed-off-by: John Hubbard <jhubbard@nvidia.com>