mm: fortify lruvec_add_folio_tail() in preparation for LRU changes
An upcoming patch will re-use the folio.lru fields and mark any pages
that do so as unevictable. In order to do so safely, all callers that
deal with LRU fields need to avoid unevictable folios.

This particular function is only protected by an internal comment. In
order to keep out any new callers that might not do the right thing,
change the comment into a VM_WARN_ON_ONCE() and return.