DIAG: instrument vmemmap_populate/free()
Although CONFIG_DEVICE_PRIVATE and hmm_range_fault() and related
functionality was first developed on x86, it also works on arm64.
However, when trying this out on an arm64 system, it turns out that
there is a massive slowdown during the setup and teardown phases.

This slowdown is due to lots of calls to WARN_ON()'s that are checking
for pages that are out of the physical range for the CPU. However,
that's a design feature of device private pages: they are specfically
chosen in order to be outside of the range of the CPU's true physical
pages. Anyway, this is all not quite correct on arm64 yet, so add some
diagnostics for now.

Signed-off-by: John Hubbard <jhubbard@nvidia.com>
cc: <stable@vger.kernel.org>