selftests: introduce and use SELFTESTS_CC_IS_CLANG instead of LLVM
Current practice in the selftests Makefiles is to use $(LLVM) as a way
to decide if clang is being used as the compiler (and/or the linker
front end). Unfortunately, this does not cover all of the use cases:

1) CC could have been set within selftests/lib.mk, by inferring it from
LLVM==1, or

2) CC could have been set externally, such as when cross compiling.

Solution: In order to allow subsystem selftests to more accurately
control clang-specific behavior, such as compiler options, provide a new
Makefile variable: SELFTESTS_CC_IS_CLANG. If $(CC) contains an
invocation of clang in any form, then SELFTESTS_CC_IS_CLANG will be
non-empty.

SELFTESTS_CC_IS_CLANG does not specify which linker is being used.
However, it can still help with linker options, because $(CC) is often
used to do both the compile and link steps (often in the same step).

Signed-off-by: John Hubbard <jhubbard@nvidia.com>