rust: transmute: reduce scope of FromBytes to Sized types
This simplifies things considerably, notably by removing the need for
the FromBytesSized proxy trait. Ideally we will want to support
non-sized types by using a derive macro, like the zerocopy crate does,
but since we are not using this trait outside of sized types at the
moment this will work fine for now.